feat(auth): add login form validators
- Rename LoginValidator.ts to login.validator.ts
- Add email and password validation rules
- Implement regex pattern for password strength
- Include French error messages
